requirements.txt
asgiref==3.5.2
Django==4.1.3
django-basicauth==0.5.3
django-cors-headers==3.13.0
django-vite==2.0.2
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
